requirements.txt
aiofiles==23.2.1
amqp==5.1.1
aniso8601==9.0.1
anyio==4.0.0
appdirs==1.4.4
apyori==1.1.2
arabic-reshaper==3.0.0
argcomplete==2.1.1
asgiref==3.4.1
asn1crypto==1.5.1
asttokens==2.2.1
backcall==0.2.0
barcode==1.0.2
bcrypt==3.2.0
beautifulsoup4==4.11.2
bidict==0.22.1
billiard==3.6.4.0
black==23.3.0
blinker==1.6.2
Brotli==1.1.0
cachelib==0.9.0
cachetools==5.3.1
captcha==0.4
celery==5.2.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
ci-info==0.3.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
config==0.5.1
configobj==5.0.8
configparser==6.0.0
contourpy==1.0.7
cryptography==40.0.2
cssselect==1.2.0
cssselect2==0.7.0
cssutils==2.7.1
cursor==1.3.4
cycler==0.11.0
decorator==5.1.1
distlib==0.3.1
Django==3.2.7
django-crispy-forms==1.13.0
dnspython==2.1.0
dominate==2.7.0
elastic-transport==8.4.0
elasticsearch==8.7.0
email-validator==1.1.3
et-xmlfile==1.1.0
etelemetry==0.3.0
exceptiongroup==1.1.3
executing==1.2.0
filelock==3.0.12
fitz==0.0.1.dev2
Flask==2.2.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Caching==2.0.2
Flask-Cors==4.0.0
Flask-Ext==0.1
Flask-JWT==0.3.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Markdown==0.3
Flask-MySQL==1.5.2
Flask-MySQLdb==0.2.0
flask-paginate==2022.1.8
Flask-RESTful==0.3.9
Flask-Session==0.4.0
Flask-Sijax==0.4.1
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.2
Flask-Uploads==0.2.1
Flask-WkHTMLtoPDF==0.1.0
Flask-WTF==1.0.0
fonttools==4.39.2
fpdf==1.7.2
frontend==0.0.3
fsspec==2023.10.0
funcy==2.0
future==0.18.3
gensim==4.3.1
gevent==23.9.0.post1
google-api-core==2.11.1
google-api-python-client==2.93.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
greenlet==2.0.1
h11==0.14.0
h5py==3.8.0
helper==2.5.0
html5lib==1.1
httplib2==0.22.0
huggingface-hub==0.19.4
idna==3.3
imbalanced-learn==0.10.1
imblearn==0.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipython==8.12.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
kombu==5.2.4
Levenshtein==0.21.1
looseversion==1.3.0
lorem-text==2.1
lxml==4.9.2
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
migrate==0.3.7
mypy-extensions==1.0.0
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.27
mysqlclient==2.0.3
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.7.0.72
openpyxl==3.1.2
oscrypto==1.3.0
packaging==23.0
pandas==1.4.0
parso==0.8.3
passlib==1.7.4
pathlib==1.0.1
pathspec==0.11.1
patsy==0.5.6
pdfkit==1.0.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
premailer==3.10.0
prompt-toolkit==3.0.38
protobuf==4.22.3
prov==2.0.0
psycopg2==2.9.6
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydot==1.4.2
pydyf==0.11.0
Pygments==2.15.0
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
PyJWT==1.4.2
pyLDAvis==3.4.0
pymongo==4.3.3
PyMuPDF==1.23.3
PyMuPDFb==1.23.3
PyMySQL==1.0.2
pyparsing==3.0.9
pypdf==3.8.1
PyPDF2==3.0.1
pyphen==0.15.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
python-barcode==0.15.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-engineio==4.7.0
python-Levenshtein==0.21.1
python-socketio==5.9.0
pytils==0.4.1
pytz==2023.3
pyxnat==1.6
PyYAML==6.0
qrcode==7.3.1
rapidfuzz==3.1.2
rdflib==7.0.0
regex==2022.10.31
reportlab==3.6.13
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.1
Sastrawi==1.0.1
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.12.2
Sijax==0.3.2
simplejson==3.19.1
six==1.15.0
sklearn==0.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==1.4.46
sqlparse==0.4.2
stack-data==0.6.2
starlette==0.31.1
statsmodels==0.14.2
stop-words==2018.7.23
svglib==1.5.1
termcolor==1.1.0
textblob==0.17.1
threadpoolctl==3.1.0
tinycss2==1.3.0
tokenizers==0.15.0
tomli==2.0.1
tools==0.1.9
tqdm==4.65.0
traitlets==5.9.0
traits==6.3.2
transformers==4.35.2
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0
uritemplate==4.1.1
uritools==4.0.1
urllib3==1.26.7
uvicorn==0.23.2
vine==5.0.0
virtualenv==20.4.3
virtualenvwrapper-win==1.2.6
visitor==0.1.3
wcwidth==0.2.6
weasyprint==62.3
webencodings==0.5.1
Werkzeug==2.2.2
wikipedia==1.4.0
wordcloud==1.9.2
wsl==0.12
WTForms==3.0.0
xhtml2pdf==0.2.11
XlsxWriter==3.1.0
yagmail==0.15.293
yellowbrick==1.5
zipp==3.11.0
zope.event==5.0
zope.interface==6.0
zopfli==0.2.3
